Share atomic file replacement
